backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
bcrypt==4.1.3
passlib>=1.7.4
tzdata>=2024.2
motor==3.3.1
pytest>=8.0.0
requests>=2.31.0
python-multipart>=0.0.9
httpx>=0.24.0
reportlab>=4.0.0
Pillow>=10.0.0
aiosmtplib==5.1.0
bleach==6.3.0
email-validator==2.3.0
slowapi==0.1.9
psutil
emergentintegrations==0.1.0
aiohttp==3.13.3
reportlab==4.4.9
aiofiles==25.1.0
